Remove -Wno-sign-compare and backport fixes from default.
--HG--
branch : release-2.x.y
extra : amend_source : 787f8b75cbfa7614b763eb98f58a7444dfdc381c
extra : histedit_source : 90b58e1814c92f4bf4185329b26d10ffd66f2c90%2C647ed17a5f1165e1c4537a6f4ff42dc30dab3b6b

+++ finch/libgnt/gntwm.c
@@ -443,7 +443,7 @@ switch_window(GntWM *wm, int direction, gboolean urgent)
 		if (pos < 0) {
 			wid = g_list_last(wm->cws->list)->data;
 			pos = g_list_length(wm->cws->list) - 1;
-		} else if (pos >= g_list_length(wm->cws->list)) {
+		} else if ((guint)pos >= g_list_length(wm->cws->list)) {
 			wid = wm->cws->list->data;
 			pos = 0;
 		} else
