+++ libpurple/stun.c
@@ -138,7 +138,7 @@ static gboolean timeoutfunc(gpointer data) {
 	sc->retry++;
 	if (sendto(sc->fd, sc->packet, sc->packetsize, 0,
 		(struct sockaddr *)&(sc->addr), sizeof(struct sockaddr_in)) !=
-		sc->packetsize)
+		(gssize)sc->packetsize)
 	{
 		purple_debug_warning("stun", "sendto failed\n");
 		return FALSE;
@@ -171,7 +171,7 @@ static void reply_cb(gpointer data, gint source, PurpleInputCondition cond) {
 	struct stun_conn *sc = data;
 	char buffer[65536];
 	char *tmp;
-	int len;
+	gssize len;
 	struct in_addr in;
 	struct stun_attrib *attrib;
 	struct stun_header *hdr;
@@ -188,13 +188,13 @@ static void reply_cb(gpointer data, gint source, PurpleInputCondition cond) {
 	}
 	buffer[len] = '\0';
 
-	if (len < sizeof(struct stun_header)) {
+	if ((gsize)len < sizeof(struct stun_header)) {
 		purple_debug_warning("stun", "got invalid response\n");
 		return;
 	}
 
 	hdr = (struct stun_header*) buffer;
-	if (len != (ntohs(hdr->len) + sizeof(struct stun_header))) {
+	if ((gsize)len != (ntohs(hdr->len) + sizeof(struct stun_header))) {
 		purple_debug_warning("stun", "got incomplete response\n");
 		return;
 	}
@@ -323,7 +323,7 @@ static void hbn_listen_cb(int fd, gpointer data) {
 
 	if(sendto(sc->fd, &hdr_data, sizeof(struct stun_header), 0,
 			(struct sockaddr *)&(sc->addr),
-			sizeof(struct sockaddr_in)) < sizeof(struct stun_header)) {
+			sizeof(struct sockaddr_in)) < (gssize)sizeof(struct stun_header)) {
 		nattype.status = PURPLE_STUN_STATUS_UNKNOWN;
 		nattype.lookup_time = time(NULL);
 		do_callbacks();
